Move MockGyro back into src since we use it as a fallback
