+++ tests/edu/wpi/first/wpilibj/MockAnalogInput.java
@@ -1,5 +1,6 @@
 package edu.wpi.first.wpilibj;
 
+import xbot.common.controls.MockRobotIO;
 import xbot.common.controls.sensors.XAnalogInput;
 
 public class MockAnalogInput implements XAnalogInput {
