+++ tests/edu/wpi/first/wpilibj/MockSolenoid.java
@@ -1,5 +1,6 @@
 package edu.wpi.first.wpilibj;
 
+import xbot.common.controls.MockRobotIO;
 import xbot.common.controls.actuators.XSolenoid;
 
 public class MockSolenoid implements XSolenoid {
