+++ tests/edu/wpi/first/wpilibj/MockSpeedController.java
@@ -2,6 +2,7 @@ package edu.wpi.first.wpilibj;
 
 import org.apache.log4j.Logger;
 
+import xbot.common.controls.MockRobotIO;
 import xbot.common.controls.actuators.XSpeedController;
 
 public class MockSpeedController implements XSpeedController {
